requirements.txt
pandas==2.2.2
Pillow==10.3.0
Requests==2.32.3
streamlit==1.32.2